Changed canvas text and cleaning
The canvas text now says: You are infected.

Introduced a separate script for handling buttons, a base one and one for the level.
With the array version for scenes, most of the work can be done by the base class
in the future. Moved scripts to level folder.
