+++ YouVsVirus/Assets/Scenes/Levelgethome/YouVsVirus_Levelgethome.unity
@@ -423,7 +423,7 @@ Transform:
   m_LocalScale: {x: 0.5, y: 0.5, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!60 &162758248
 PolygonCollider2D:
@@ -783,7 +783,7 @@ RectTransform:
   - {fileID: 111896179}
   - {fileID: 1147243145}
   m_Father: {fileID: 0}
-  m_RootOrder: 10
+  m_RootOrder: 11
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -872,7 +872,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1293853906}
+      - m_Target: {fileID: 801701337}
         m_MethodName: Continue
         m_Mode: 1
         m_Arguments:
@@ -1119,7 +1119,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1293853906}
+      - m_Target: {fileID: 801701337}
         m_MethodName: Retry
         m_Mode: 1
         m_Arguments:
@@ -1523,6 +1523,49 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 761891737}
   m_CullTransparentMesh: 0
+--- !u!1 &801701336
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 801701338}
+  - component: {fileID: 801701337}
+  m_Layer: 0
+  m_Name: ButtonHandle
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &801701337
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 801701336}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 11d4ec839c429a64b949f718472f9474, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!4 &801701338
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 801701336}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 9.981266, y: -5.06858, z: -3795.1484}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 6
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &934205702
 GameObject:
   m_ObjectHideFlags: 0
@@ -1654,7 +1697,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1293853906}
+      - m_Target: {fileID: 801701337}
         m_MethodName: MainMenu
         m_Mode: 1
         m_Arguments:
@@ -1807,7 +1850,7 @@ Transform:
   m_LocalScale: {x: 0.5, y: 0.5, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1147243144
 GameObject:
@@ -1895,6 +1938,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1293853907}
   - component: {fileID: 1293853906}
+  - component: {fileID: 1293853908}
   m_Layer: 0
   m_Name: EndLevelController
   m_TagString: Untagged
@@ -1914,9 +1958,14 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1ae52b6270c0ad148a94f5770b491669, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  startTime: 0
   LevelTimeout: 0
   EndConditionMetDelay: 3
+  playerDied: 0
   playerExposed: 0
+  activeInfections: 0
+  endConditionMet: 0
+  playerHome: 0
   CanvasFail: {fileID: 2046121199}
   CanvasSucc: {fileID: 236133598}
   CreateHumans: {fileID: 742637914}
@@ -1934,6 +1983,24 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1293853908
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1293853905}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9465a36d45b26d44d85090146c1b28e8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  LevelTimeout: 0
+  EndConditionMetDelay: 3
+  playerExposed: 0
+  CanvasFail: {fileID: 2046121199}
+  CanvasSucc: {fileID: 236133598}
+  CreateHumans: {fileID: 742637914}
 --- !u!1 &1342456105
 GameObject:
   m_ObjectHideFlags: 0
@@ -2092,7 +2159,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1293853906}
+      - m_Target: {fileID: 801701337}
         m_MethodName: MainMenu
         m_Mode: 1
         m_Arguments:
@@ -2248,7 +2315,7 @@ RectTransform:
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
   m_AnchoredPosition: {x: 0, y: 147}
-  m_SizeDelta: {x: 300, y: 50}
+  m_SizeDelta: {x: 400, y: 50}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1959990384
 MonoBehaviour:
@@ -2271,7 +2338,7 @@ MonoBehaviour:
   m_text: 'Sorry.
 
 
-    You failed.'
+    You are infected.'
   m_isRightToLeft: 0
   m_fontAsset: {fileID: 11400000, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
   m_sharedMaterial: {fileID: 2180264, guid: 8f586378b4e144a9851e7b34d9b748ee, type: 2}
@@ -2340,10 +2407,10 @@ MonoBehaviour:
   m_margin: {x: 0, y: 0, z: 0, w: 0}
   m_textInfo:
     textComponent: {fileID: 1959990384}
-    characterCount: 19
+    characterCount: 25
     spriteCount: 0
-    spaceCount: 3
-    wordCount: 3
+    spaceCount: 4
+    wordCount: 4
     linkCount: 0
     lineCount: 3
     pageCount: 1
@@ -2545,7 +2612,7 @@ RectTransform:
   - {fileID: 168650350}
   - {fileID: 1342456106}
   m_Father: {fileID: 0}
-  m_RootOrder: 9
+  m_RootOrder: 10
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -2736,7 +2803,7 @@ Transform:
   m_LocalScale: {x: 0.049455, y: 0.049455, z: 1.1093745}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &2129108443
 SpriteRenderer:
